+++ tests/Instances.hs
@@ -155,6 +155,14 @@ instance Arbitrary a => Arbitrary (SomeType a) where
                       , Record  <$> arbitrary <*> arbitrary <*> arbitrary
                       ]
 
+instance Arbitrary EitherTextInt where
+    arbitrary = oneof
+        [ LeftBool <$> arbitrary
+        , RightInt <$> arbitrary
+        , BothTextInt <$> arbitrary <*> arbitrary
+        , pure NoneNullary
+        ]
+
 instance Arbitrary (GADT String) where
     arbitrary = GADT <$> arbitrary